requirements.txt
Flask==1.1.1
gunicorn==19.9.0
line-bot-sdk==1.14.0
APScheduler==3.6.1
psycopg2==2.8.3
redis==3.4.1
selenium==3.8.0
datetime
